+++ sdk/resourcemanager/servicelinker/armservicelinker/options.go
@@ -14,12 +14,6 @@ type LinkerClientBeginCreateOrUpdateOptions struct {
 	ResumeToken string
 }
 
-// LinkerClientBeginDeleteOptions contains the optional parameters for the LinkerClient.BeginDelete method.
-type LinkerClientBeginDeleteOptions struct {
-	// Resumes the LRO from the provided token.
-	ResumeToken string
-}
-
 // LinkerClientBeginUpdateOptions contains the optional parameters for the LinkerClient.BeginUpdate method.
 type LinkerClientBeginUpdateOptions struct {
 	// Resumes the LRO from the provided token.
@@ -39,11 +33,6 @@ type LinkerClientGetOptions struct {
 
 // LinkerClientListConfigurationsOptions contains the optional parameters for the LinkerClient.ListConfigurations method.
 type LinkerClientListConfigurationsOptions struct {
-	// placeholder for future optional parameters
-}
-
-// LinkerClientListOptions contains the optional parameters for the LinkerClient.NewListPager method.
-type LinkerClientListOptions struct {
 	// placeholder for future optional parameters
 }