+++ sdk/resourcemanager/servicelinker/armservicelinker/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Release History
 
+## 2.0.0 (2024-03-14)
+### Breaking Changes
+
+- Function `*LinkerClient.BeginDelete` has been removed
+- Function `*LinkerClient.NewListPager` has been removed
+
+### Features Added
+
+- New value `ClientTypeKafkaSpringBoot` added to enum type `ClientType`
+
+
 ## 1.2.0 (2023-11-24)
 ### Features Added